EMSUSD-1111 move all transform-related UFE code to their own folder
This make finding all transform-related code easier.
Also, remove a few unused classes.
