+++ lib/mayaUsd/ufe/trf/RotationUtils.cpp
