+++ lib/mayaUsd/ufe/trf/RotationUtils.h
