+++ lib/mayaUsd/ufe/trf/UsdPointInstanceModifierBase.h
