+++ lib/mayaUsd/ufe/trf/UsdPointInstanceOrientationModifier.cpp
