+++ lib/mayaUsd/ufe/trf/UsdPointInstancePositionModifier.cpp
