+++ lib/mayaUsd/ufe/trf/UsdPointInstanceScaleModifier.cpp
