+++ lib/mayaUsd/ufe/trf/UsdTransform3dBase.cpp
