+++ lib/mayaUsd/ufe/trf/UsdTransform3dPointInstance.h
@@ -17,10 +17,10 @@
 #define MAYAUSD_UFE_USD_TRANSFORM3D_POINT_INSTANCE_H
 
 #include <mayaUsd/base/api.h>
-#include <mayaUsd/ufe/UsdPointInstanceOrientationModifier.h>
-#include <mayaUsd/ufe/UsdPointInstancePositionModifier.h>
-#include <mayaUsd/ufe/UsdPointInstanceScaleModifier.h>
-#include <mayaUsd/ufe/UsdTransform3dBase.h>
+#include <mayaUsd/ufe/trf/UsdPointInstanceOrientationModifier.h>
+#include <mayaUsd/ufe/trf/UsdPointInstancePositionModifier.h>
+#include <mayaUsd/ufe/trf/UsdPointInstanceScaleModifier.h>
+#include <mayaUsd/ufe/trf/UsdTransform3dBase.h>
 
 #include <usdUfe/ufe/UsdSceneItem.h>
