+++ lib/mayaUsd/ufe/trf/UsdTransform3dRead.cpp
