+++ lib/mayaUsd/ufe/trf/UsdTransform3dReadImpl.cpp
