+++ lib/mayaUsd/ufe/trf/XformOpUtils.cpp
