+++ lib/mayaUsd/ufe/trf/XformOpUtils.h
